Initial commit containing basic objects
Limited functionality right now. Constructors include: 
XYPair, which contains an x and a y property, and an add function, to add the x and y of another object. Used by Position and Vector.

Position, which for now is an XYPair by another name. Used by Shape

Vector, contains the x and y components of some vector. Contains methods to get the angle and magnitude, as well as manipulate the x and y components. Used by Movable

Shape, kind of an abstraction. It adds methods that all shapes should have, such as a position, and the methods to modify said position. All shapes call this constructor.

Movable, which adds a vector to some object that is movable. If the object has no position, the object cannot be made movable. It also contains methods to set the velocity, and a move function, which adds the vector's x and y components to the object's position.

Rect, which represents a rectangle, with a width and height. Contains a draw function which takes the canvas it will be drawn to, and draws the rectangle.

Circle, which represents a circle. Also contains a draw function to draw on a given canvas.